[PLAYER-4778] move autofocus to utils folder
